Update requirements and logging; add new ETL paths and enhance data processing
