+++ requirements.txt
@@ -13,4 +13,5 @@ wget==3.2
 openpyxl==3.1.5
 pyarrow==19.0.1
 pydeck==0.9.1
+keplergl==0.3.2
 unidecode==1.3.8
