+++ src/config.json
@@ -23,7 +23,7 @@
         "ftp_url": "/Censos/Censo_Demografico_2022/Agregados_por_Setores_Censitarios/Agregados_por_Setor_csv/",
         "format": "zip",
         "file_name": "Agregados_por_setores_basico_BR.zip",
-        "raw_path": "data/raw/ibge/sociodemographics"
+        "raw_path": "data/landing_area/ibge/sociodemographics"
       },
       {
         "name": "IBGE - SCs",
@@ -32,7 +32,7 @@
         "ftp_url": "/organizacao_do_territorio/malhas_territoriais/malhas_de_setores_censitarios__divisoes_intramunicipais/censo_2022/setores/gpkg/BR/",
         "format": "gpkg",
         "file_name": "BR_setores_CD2022.gpkg",
-        "raw_path": "data/raw/ibge/scs"
+        "raw_path": "data/landing_area/ibge/scs"
       }
     ],
     "sspsp": [
@@ -42,7 +42,7 @@
         "years": [2024,2023,2022,2021,2020,2019,2018,2017],
         "url": "https://www.ssp.sp.gov.br/assets/estatistica/transparencia/baseDados/celularesSub/CelularesSubtraidos_{}.xlsx",
         "format": "xlsx",
-        "raw_path": "data/raw/criminality/cellphones"
+        "raw_path": "data/landing_area/sspsp/cellphones"
       },
       {
         "name": "Criminality - Vehicles",
@@ -50,7 +50,7 @@
         "years": [2024,2023,2022,2021,2020,2019,2018,2017],
         "url": "https://www.ssp.sp.gov.br/assets/estatistica/transparencia/baseDados/veiculosSub/VeiculosSubtraidos_{}.xlsx",
         "format": "xlsx",
-        "raw_path": "data/raw/criminality/vehicles"
+        "raw_path": "data/landing_area/sspsp/vehicles"
       },
       {
         "name": "Criminality - Other Crimes",
@@ -58,7 +58,7 @@
         "years": [2024,2023,2022],
         "url": "https://www.ssp.sp.gov.br/assets/estatistica/transparencia/spDados/SPDadosCriminais_{}.xlsx",
         "format": "xlsx",
-        "raw_path": "data/raw/criminality/other_crimes"
+        "raw_path": "data/landing_area/sspsp/criminality/other_crimes"
       }
     ]
   },
