+++ src/logs/20250319_214839_EVALUATION.log
@@ -1,0 +1,2 @@
+2025-03-19 21:53:23 - EVALUATION - INFO - dissolve_sedona: 261.6176338195801s
+2025-03-19 22:03:25 - EVALUATION - INFO - dissolve_geopandas: 601.6481568813324s
