+++ src/logs/monitoring/2025-03-12_etl.csv
@@ -1,0 +1,35 @@
+table,level,read_time,treat_time,write_time,test_time,row_count,total_time
+bronze_crimes_cellphones/2024,BRONZE,174.27,0.53,17.32,2.44,412527,192.12
+bronze_crimes_cellphones/2023,BRONZE,204.54,0.33,13.83,1.91,902681,218.70000000000002
+bronze_crimes_cellphones/2022,BRONZE,155.85,0.29,11.52,1.87,1345411,167.66
+bronze_crimes_cellphones/2021,BRONZE,135.48,0.28,10.14,2.04,1737655,145.89999999999998
+bronze_crimes_cellphones/2020,BRONZE,130.76,0.28,10.41,2.37,2110690,141.45
+bronze_crimes_cellphones/2019,BRONZE,179.44,0.22,13.91,3.15,2629015,193.57
+bronze_crimes_cellphones/2018,BRONZE,171.36,0.28,12.5,3.0,3122601,184.14000000000001
+bronze_crimes_cellphones/2017,BRONZE,181.16,0.25,13.35,3.16,3644312,194.76
+bronze_crimes_vehicles/2024,BRONZE,113.46,0.28,8.55,0.7,274733,122.28999999999999
+bronze_crimes_vehicles/2023,BRONZE,123.08,0.27,8.84,0.92,583074,132.19
+bronze_crimes_vehicles/2022,BRONZE,83.42,0.22,7.61,1.15,799262,91.25
+bronze_crimes_vehicles/2021,BRONZE,66.24,0.25,6.77,1.26,973354,73.25999999999999
+bronze_crimes_vehicles/2020,BRONZE,61.45,0.25,6.01,1.58,1136603,67.71000000000001
+bronze_crimes_vehicles/2019,BRONZE,85.78,0.25,7.62,1.48,1363930,93.65
+bronze_crimes_vehicles/2018,BRONZE,99.71,0.24,8.41,1.86,1628499,108.35999999999999
+bronze_crimes_vehicles/2017,BRONZE,103.93,0.25,8.11,1.99,1905879,112.29
+bronze_crimes_others/2024,BRONZE,274.61,0.15,17.1,1.87,1191801,291.86
+bronze_crimes_others/2023,BRONZE,289.25,0.13,20.92,2.83,2433710,310.3
+bronze_crimes_others/2022,BRONZE,246.6,0.13,17.73,4.18,3631683,264.46
+bronze_scs_demographics/2022,BRONZE,0.43,0.24,9.24,1.16,468099,9.91
+bronze_scs_geom/2022,BRONZE,0.16,0.19,133.77,49.78,472780,134.12
+silver_scs_demographics,SILVER,0.09,0.0,2.42,0.88,103319,3.42
+silver_mun_demographics,SILVER,0.11,0.0,0.78,0.51,645,1.45
+silver_scs_geom,SILVER,0.1,0.0,10.81,7.06,103319,18.19
+silver_mun_geom,SILVER,0.08,0.0,32.67,1.09,645,33.91
+silver_scs,SILVER,0.13,0.0,1.92,6.89,103319,8.98
+silver_mun,SILVER,0.16,0.0,4.35,6.31,645,10.87
+silver_crimes_cellphones,SILVER,0.12,0.36,46.74,5.02,2239322,52.24
+silver_crimes_vehicles,SILVER,0.12,0.26,19.91,2.34,1097697,22.63
+silver_crimes_others,SILVER,0.1,0.25,34.05,4.09,2560725,38.5
+silver_crimes,SILVER,0.3,0.0,313.79,95.31,5047988,409.63
+gold_crimes,GOLD,2.25,0.0,44.45,4.78,5047988,51.48
+gold_crimes_scs,GOLD,2.25,0.0,58.62,3.4,532229,80.21
+gold_crimes_mun,GOLD,3.2699999999999996,0.0,43.48,1.72,4952,63.11
